start user with cocktails and menu

+++ ShakeitServer/Models/CocktailIngredient.cs
@@ -9,6 +9,6 @@ namespace ShakeitServer.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Pour { get; set; }
+        public decimal Pour { get; set; }
     }
 }
